% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_savant_games.R
\name{available_savant_games}
\alias{available_savant_games}
\title{**Get Available College/Summer League Games in Baseball Savant**}
\usage{
available_savant_games(level = "College Baseball")
}
\arguments{
\item{level}{The level to select from. Available levels: College Baseball, MLB Draft League, Appalachian League,
Cape Cod Baseball League (NO STATCAST DATA AVAILABLE FOR Cape Cod)}
}
\value{
A data frame containing the game date, the game pk, away team, home team, venue, league name, play by play data available,
statcast data available
}
\description{
Title
}
\examples{
\donttest{
  try(available_savant_games(level = "College Baseball"))
  try(available_savant_games(level = "MLB Draft League"))
  try(available_savant_games(level = "Appalachian League"))
  try(available_savant_games(level = "Cape Cod Baseball League"))
}
}
