% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_schedule.R
\name{ncaa_schedule}
\alias{ncaa_schedule}
\title{**Get Schedule and Results for NCAA Baseball Teams**}
\usage{
ncaa_schedule(team_id = NULL, year = NULL)
}
\arguments{
\item{team_id}{The team's new unique NCAA id}

\item{year}{The season i.e. use 2024 for the 2023-2024 season, etc.}
}
\value{
A data frame that returns a teams schedule info along with both the home and away team ids, conference, and division, as well as
the result of the game, innings, game_info_url, and a contest_id
}
\description{
Title
}
\examples{

 try(ncaa_schedule_info(team_id = 574226, year = 2024))

}
