% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_teams.R
\name{ncaa_teams}
\alias{ncaa_teams}
\title{**Get NCAA Baseball Teams**}
\usage{
ncaa_teams(years = 2024, divisions = c(1, 2, 3))
}
\arguments{
\item{years}{A year or range of years to acquire data (2013 is the earliest)}

\item{divisions}{One of three (or all 3 options) divisions: 1,2, and/or 3}
}
\value{
A data frame containing each team's new id, the year, the team_name, the conference + conference_id, and season + season_id
}
\description{
Title
}
\examples{
\donttest{
  try(ncaa_teams(years = c(2013:2024), divisions = c(1,2,3)))
  try(ncaa_teams(years = 2024, divisions = 1))
}
}
