% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast_pbp_college.R
\name{statcast_pbp_college}
\alias{statcast_pbp_college}
\title{**Acquire Statcast or raw play-by-play data from collegiate/summer league games**}
\usage{
statcast_pbp_college(game_pk)
}
\arguments{
\item{game_pk}{The game pk acquired from running the available_savant_games function}
}
\value{
A data frame containing the following columns:

colname  type
play_id: character
inning: numeric
ab_number: numeric
cap_index: numeric
outs: numeric
batter: numeric
stand: character
batter_name: character
pitcher: numeric
p_throws: character
pitcher_name: character
team_batting: character
team_fielding: character
team_batting_id: numeric
team_fielding_id: numeric
result: character
des: character
events: character
strikes: numeric
balls: numeric
pre_strikes: numeric
pre_balls: numeric
call: character
call_name: character
description: character
result_code: character
pitch_call: character
is_strike_swinging: logical
balls_and_strikes: character
sz_top: numeric
sz_bot: numeric
extension: numeric
plateTime: numeric
zone: numeric
spin_rate: numeric
px: numeric
pz: numeric
x0: numeric
y0: numeric
z0: numeric
ax: numeric
ay: numeric
az: numeric
vx0: numeric
vy0: numeric
vz0: numeric
pfxX: numeric
pfxZ: numeric
pfxZWithGravity: numeric
pfxZWithGravityNice: numeric
pfxZDirection: character
pfxXWithGravity: numeric
pfxXNoAbs: numeric
pfxXDirection: character
breakX: numeric
breakZ: numeric
inducedBreakZ: numeric
is_bip_out: character
pitch_number: numeric
player_total_pitches: numeric
player_total_pitches_pitch_types: numeric
game_total_pitches: numeric
rowId: character
game_pk: character
player_name: character
hc_x: numeric
hc_x_ft: numeric
hc_y: numeric
hc_y_ft: numeric
runnerOn1B: logical
runnerOn2B: logical
runnerOn3B: logical
}
\description{
**Acquire Statcast or raw play-by-play data from collegiate/summer league games**
}
\examples{
\donttest{
  try(statcast_pbp_college(769400))
}
}
